Rename reducer for consistency
And rename local params and vars.

See previous commit for equivalent justification.
