+++ packages/harpguru-core/src/components/menu-of-covariants/utils/reduce-covariants-to-harpstrata/index.ts
