--- packages/harpguru-core/src/components/menu-of-covariants/utils/reduce-for-new-harp-strata-by-covariants/index.ts
@@ -1,1 +1,0 @@
-export { reduceForNewHarpStrataByCovariants } from './reduce-for-new-harp-strata-by-covariants'
